make middleware and gate

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('email')->unique()->nullable();
             $table->string('email_verified_at')->nullable();
             $table->string('google_id')->nullable();
-            $table->unsignedBigInteger('level_id')->nullable();
+            $table->integer('level_id')->nullable();
             $table->string('password', 100)->nullable();
 
             $table->timestamps();
